feat(frontend): Add ability to compare 2 maps
This is the first pass at adding the ability to compare 2 maps. There
are some hard-coded values I don't like and the code needs a bit of a
tidy-up, but it does work. We lay 2 maps one on top the other and clip
the div size based on where the drag handler position is horizontally.