feat(frontend): Add ability to bootstrap cient
Added the ability to fetch data, set/interact with feature flags etc.
Basically, I'd like to see this as a general solution to local development
where you need to put the client into a known state, before you test
your situation. This can take a lot of time, especially if your constantly
changing code and re-loading, then having to reset to the known state,
so doing this automatically would be a real time saver.